finalizando index
