+++ FrontCaixa/uix/kv/macro/sw_table_with_title.kv
@@ -1,0 +1,26 @@
+<SWTableWithTitle>:
+    orientation:'vertical'
+    adaptive_width:True
+    spacing:dp(5)
+    pos_hint:{'top':1}
+    MDBoxLayout:
+        orientation:'vertical'
+        radius:dp(7),dp(7),dp(7),dp(7)
+        md_bg_color:24/255,24/255,35/255,1
+        adaptive_width:True
+        padding:dp(4)
+        MDBoxLayout:
+            size_hint_y:None
+            height:dp(35)
+            MDLabel:
+                text:'Produtos Adicionados'
+                font_style:'Overline'
+                font_size:dp(22)
+                halign:'center'
+                bold:True
+                color:1,1,1,1
+        MDBoxLayout:
+            id:table
+            size_hint_x:None
+            md_bg_color:1,1,1,1
+            width:dp(690)
