+++ FrontCaixa/uix/kv/micro/sw_input_with_title.kv
@@ -1,0 +1,33 @@
+<SWInputWithTitle>:
+    adaptive_height:True
+    md_bg_color:24/255,24/255,35/255,1
+    radius:dp(7),dp(7),dp(7),dp(7)
+    padding:dp(4)
+    orientation:'vertical'
+    MDBoxLayout:
+        size_hint_y:None
+        height:dp(35)
+        MDLabel:
+            text:root.title
+            font_style:'Overline'
+            font_size:dp(22)
+            halign:'center'
+            bold:True
+            color:1,1,1,1
+    MDBoxLayout:
+        md_bg_color:1,1,1,1
+        radius:dp(7),dp(7),dp(7),dp(7)
+        size_hint_y:None
+        height:dp(35)
+        padding:dp(5),dp(0)
+        TextInput:
+            id:total_recebido
+            text:root.text
+            font_style:'Overline'
+            font_size:dp(20)
+            halign:'right'
+            multiline:False
+            write_tab:False
+            input_filter:'float'
+            background_color:0,0,0,0
+            on_text:root.set_text(self.text)
