+++ FrontCaixa/uix/kv/micro/sw_label_with_title.kv
@@ -1,0 +1,27 @@
+<SWLabelWithTitle>:
+    adaptive_height:True
+    md_bg_color:24/255,24/255,35/255,1
+    radius:dp(7),dp(7),dp(7),dp(7)
+    padding:dp(4)
+    orientation:'vertical'
+    MDBoxLayout:
+        size_hint_y:None
+        height:dp(35)
+        MDLabel:
+            text:root.title
+            font_style:'Overline'
+            font_size:dp(22)
+            halign:'center'
+            bold:True
+            color:1,1,1,1
+    MDBoxLayout:
+        md_bg_color:1,1,1,1
+        radius:dp(7),dp(7),dp(7),dp(7)
+        size_hint_y:None
+        height:dp(35)
+        padding:dp(5),dp(0)
+        MDLabel:
+            text:root.text_start+root.text+root.text_end
+            font_style:'Overline'
+            font_size:dp(22)
+            halign:'right'
