+++ FrontCaixa/uix/object/micro/sw_input_with_title.py
@@ -1,0 +1,7 @@
+from kivymd.uix.boxlayout import MDBoxLayout
+from kivy.properties import StringProperty
+class SWInputWithTitle(MDBoxLayout):
+    title = StringProperty('')
+    text = StringProperty('')
+    def set_text(self,text):
+        self.text = text
