+++ FrontCaixa/uix/object/micro/sw_label_with_title.py
@@ -1,0 +1,8 @@
+from kivymd.uix.boxlayout import MDBoxLayout
+from kivy.properties import StringProperty
+class SWLabelWithTitle(MDBoxLayout):
+    title = StringProperty('')
+    text = StringProperty('')
+    text_start = StringProperty('')
+    text_end = StringProperty('')
+
